lexer and word

+++ src/Float_c.java
@@ -1,6 +1,6 @@
 public class Float_c extends Token {
     double value;
-    public Float_c(int value) {
+    public Float_c(double value) {
         super(Tag.FLOAT_C);
         this.value = value;
     }
